+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 sine phase. This makes a Click value of 1.0 more meaningful, as "full click".
 * Change: Improved knob labels and help texts.
 * Change: Improved explanation of F1 help mode.
+* Feature: Added a filter implementation for filtering the output.
 
 
 0.2.1 (2021-01-09)